pyproject.toml
[tool.poetry]
name = "notes"
version = "0.1.0"
description = ""
authors = ["Mamatkul <mamatkul1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
pydantic = {extras = ["email"], version = "^2.5.1"}
bcrypt = "^4.0.1"
uvicorn = "^0.24.0.post1"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
alembic = "^1.12.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
sqlalchemy-stubs = "^0.4"
flake8 = "^6.0.0"
autoflake = "^2.2.0"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
